Respect libcxx setting and set language standard.
Ensure that the libcxx setting actually sets the _GLIBCXX_USE_CXX11_ABI
macro in order to control the ABI in use. Also set the C++ language
standard in use to the most recent supported by the compiler, as
boost can hopefully take advantage (or at the very least cope with) a
recent language standard.